third_party/blink/web_tests/external/wpt/css/css-forms/appearance-base-basic-expected.txt

This is a testharness.js-based test.
[FAIL] Basic appearance stylesheet properties for input-button
  assert_equals: color should inherit expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
[FAIL] Basic appearance stylesheet properties for input-reset
  assert_equals: color should inherit expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
[FAIL] Basic appearance stylesheet properties for button
  assert_equals: color should inherit expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
